requirements.txt
Django==5.1.4
djangorestframework==3.15.2
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
pgvector==0.3.6
boto3==1.35.36
google-genai==1.26.0
dodopayments==1.70.0
whitenoise==6.7.0
python-slugify==8.0.4
gunicorn==22.0.0
